% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Distribution par jour}
\usage{
calcul_distribution_semaine(df)
}
\arguments{
\item{df}{Un data.frame.}
}
\value{
Un data.frame.
}
\description{
Distribution par jour
}
