% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_boucle}
\alias{compter_nombre_boucle}
\title{Compter le nombre de boucles}
\usage{
compter_nombre_boucle(df)
}
\arguments{
\item{df}{Un data.frame.}
}
\value{
Un entier.
}
\description{
Compter le nombre de boucles
}
