% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_velo}
\alias{df_velo}
\title{Donnees de trajets velo (Toussaint 2025)}
\format{
Un data.frame
}
\source{
Donnees du cours
}
\usage{
df_velo
}
\description{
Jeu de donnees contenant les trajets velo pendant les vacances
de la Toussaint 2025.
}
\keyword{datasets}
