% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtrer les anomalies}
\usage{
filtre_anomalie(df)
}
\arguments{
\item{df}{Un data.frame contenant les trajets vélos avec les colonnes suivantes :
\describe{
\item{Probabilité.de.présence.d.anomalies}{Niveau d'anomalie ("Faible", "Forte", etc.)}
}}
}
\value{
Un data.frame sans les observations présentant des anomalies faibles ou fortes.
}
\description{
Cette fonction filtre les lignes correspondant aux anomalies faibles ou fortes
dans un jeu de données de trajets vélo.
}
