% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer les trajets par boucle}
\usage{
filtrer_trajet(df, boucles)
}
\arguments{
\item{df}{Un data.frame.}

\item{boucles}{Un vecteur de numeros de boucle.}
}
\value{
Un data.frame filtre.
}
\description{
Filtrer les trajets par boucle
}
