% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver le trajet maximum}
\usage{
trouver_trajet_max(df)
}
\arguments{
\item{df}{Un data.frame.}
}
\value{
Un data.frame.
}
\description{
Trouver le trajet maximum
}
