% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_boucle}
\alias{compter_nombre_boucle}
\title{Compter le nombre de boucles de comptage}
\usage{
compter_nombre_boucle(df)
}
\arguments{
\item{df}{Un data.frame contenant les trajets vélos avec les colonnes suivantes :
\describe{
\item{Boucle.de.comptage}{Identifiant ou nom de la boucle de comptage}
}}
}
\value{
Un entier correspondant au nombre de boucles uniques.
}
\description{
Cette fonction calcule le nombre de boucles de comptage distinctes.
}
