% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_trajets}
\alias{compter_nombre_trajets}
\title{Compter le nombre de trajets}
\usage{
compter_nombre_trajets(df)
}
\arguments{
\item{df}{Un data.frame.}
}
\value{
Un entier.
}
\description{
Compter le nombre de trajets
}
