% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtrer les anomalies}
\usage{
filtre_anomalie(df)
}
\arguments{
\item{df}{Un data.frame.}
}
\value{
Un data.frame filtre.
}
\description{
Supprime les lignes avec anomalies faibles ou fortes.
}
