% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Plot distribution}
\usage{
plot_distribution_semaine(df)
}
\arguments{
\item{df}{Un data.frame.}
}
\value{
Un graphique.
}
\description{
Plot distribution
}
